requirements.txt
streamlit
azure-ai-inference
azure-core